move astria config to separate file, add some validation
